+++ regen/objects/sky/sky.h
@@ -4,7 +4,7 @@
 #include <list>
 
 #include <regen/camera/camera.h>
-#include <regen/states/light-state.h>
+#include <regen/shading/light-state.h>
 #include <regen/objects/sky/sky-layer.h>
 #include <regen/objects/primitives/rectangle.h>
