+++ regen/scene/state-configurer.cpp
@@ -1,4 +1,4 @@
-#include <regen/states/light-state.h>
+#include <regen/shading/light-state.h>
 #include <regen/objects/mesh.h>
 #include <regen/utility/string-util.h>
