+++ regen/shading/direct-shading.cpp
