+++ regen/shading/direct-shading.h
@@ -9,7 +9,7 @@
 #define __SHADING_DIRECT_H_
 
 #include <regen/scene/state.h>
-#include <regen/states/light-state.h>
+#include <regen/shading/light-state.h>
 #include <regen/textures/texture-state.h>
 #include <regen/camera/light-camera.h>
