+++ regen/shading/light-pass.cpp
