+++ regen/shading/light-pass.h
