+++ regen/shading/light-state.cpp
