+++ regen/shading/light-state.h
